Refactor: Move element_has_css_class to unittest_base.
